pyproject.toml
[tool.poetry]
name = "OpenModelicaCompiler"
version = "0.3.0a0"
description = "OpenModelica compiler (omc) interface for Python"
repository = "https://github.com/ijknabla/OpenModelicaCompilerForPython"
authors = ["ijknabla <ijknabla@gmail.com>"]
readme = "README.md"
license = "GPL-3.0, OSMC-PL 1.2 (user's choice)"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: User Interfaces",
]
packages = [
    { include = "omc4py"},
]

[tool.poetry.dependencies]
python = "^3.7"
Arpeggio = ">=1.6"
exceptiongroup = ">=1.0"
pyzmq = ">=17.0.0"
typing-extensions = ">=3.7.4.2"
modelicalang = "^0.1.0a0"

[tool.poetry.group.bootstrap.dependencies]
click = "^8.1.3"
pyyaml = "^6.0"
tqdm = "^4.62.3"

[tool.poetry.group.dev.dependencies]
isort = "^5.8"
black = ">=23"

[tool.poetry.group.lint.dependencies]
mypy = "^1"
pyproject-flake8 = "^0.0.1-alpha.2"
types-arpeggio = {git = "https://github.com/ijknabla/types-arpeggio.git", rev = "main"}
types-pyyaml = "^6.0.12.10"
types-setuptools = "^65.7.0.3"
types-tqdm = "^4.65.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.20"
pytest-cov = "^4.0.0"
pytest-dependency = "^0.5.1"
typing-extensions = ">=4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.coverage.report]
exclude_lines = [
    "^\\s*class \\w+\\(Protocol",
    "^\\s*@overload$",
    "^\\s*if TYPE_CHECKING",
    "^\\s*raise NotImplementedError",
    "#\\s*pragma\\s*:\\s*no\\s*cover",
]

[tool.flake8]
exclude = ".venv"

# Allow too-long line in auto-generated modules
per-file-ignores = """
    omc4py/enumeration.py:E501,
    omc4py/v_*_*/*.py:E501,
"""

[tool.isort]
profile = "black"
line_length = 79

[[tool.mypy.overrides]]
module = [
    "numpy",
]
ignore_missing_imports = true
